+++ Sources/SwiftSDK/Messaging/RT/Channel.swift
@@ -121,4 +121,17 @@
         removeCommandListeners()
         removeUserStatusListeners()
     }
+    
+    open func sendCommand(commandType: String, data: Any?, responseHandler: (() -> Void)!, errorHandler: ((Fault) -> Void)!) {
+        let wrappedBlock: (Any) -> () = { response in
+            responseHandler()
+        }        
+        if let channelName = self.channelName {
+            var options = ["channel": channelName, "type": commandType] as [String : Any]
+            if let data = data {
+                options["data"] = JSONUtils.shared.objectToJSON(objectToParse: data)
+            }
+            RTMethod.shared.sendCommand(type: PUB_SUB_COMMAND, options: options, responseHandler: wrappedBlock, errorHandler: errorHandler)
+        }
+    }
 }